added deleteBeerById not found test and changed signature to return boolean

+++ src/main/java/guru/springframework/yudi/spring6restmvcyudi/services/BeerServiceJPA.java
@@ -80,10 +80,17 @@ public class BeerServiceJPA implements BeerService {
 
     /**
      * @param beerId
+     * @return
      */
     @Override
-    public void deleteById(UUID beerId) {
-        beerRepository.deleteById(beerId);
+    public Boolean deleteById(UUID beerId) {
+        if(beerRepository.existsById(beerId)) {
+            beerRepository.deleteById(beerId);
+            return true;
+        }
+        else{
+            return false;
+        }
     }
 
     /**
